osx: clean up all children with waitpid whenever the timeout expires
Previously, it would only clean up a single child every 10ms, which made
it so that, if a test ran fast enough, it would be the only one that got
cleaned up, and it would thus be the only test that would run. This makes
sure that all finished tests are always cleaned.